+++ CHANGELOG.md
@@ -16,6 +16,11 @@ Types of Changes:
 
 ## [Unreleased]
 
+### Added
+
+-   POST `/api/hackers/resume/` endpoint for uploading a hacker's resume before the hacker is actually created in the DB. Unattached resumes are deleted after 24 hours.
+-   Added the ability to upload the hacker document as a `application/json` while providing the optional `resume_id` value to attach a resume to a hacker document which is returned from the `/api/hackers/resume/` endpoint.
+
 ### Security
 
 -   Removed ability for user to provide redirect uri for email verification endpoint.